requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
crispy-bootstrap4==2024.10
crispy-bootstrap5==2024.10
cryptography==44.0.0
dj-database-url==2.3.0
Django==5.1.4
django-bootstrap4==24.4
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-email==0.1.10
django-environ==0.11.2
django-filter==24.3
django-phonenumber-field==8.0.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
environ==1.0
et_xmlfile==2.0.0
idna==3.10
Markdown==3.7
numpy==2.2.2
openpyxl==3.1.5
pandas==2.2.3
paypalrestsdk==1.13.3
phonenumbers==8.13.53
pillow==11.0.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
reportlab==4.2.5
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
whitenoise==6.8.2
